+++ script/module/html.js
@@ -1,8 +1,11 @@
 /* 在 HTML 块中使用的小工具 */
 
+/* 运行系统命令 */
+const child_process = require('child_process');
+
 function runcmd(commands) {
     if (window.confirm(commands)) {
         commands = `start powershell -c ${commands.replaceAll('\n', '; ')}pause`;
-        require('child_process').exec(commands, null);
+        child_process.exec(commands, null);
     }
 }